% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdata.R
\docType{data}
\name{diffdata}
\alias{diffdata}
\title{Diff Data}
\format{
A data frame with 30 rows and 3 variables
\describe{
\item{record_id}{The ID variable of each subject}
\item{timepoint}{The timepoint of each data collection event, such as year}
\item{test_score}{The test score, a continuous variable}
}
}
\usage{
diffdata
}
\description{
Test data for testing longdiff function
}
\keyword{datasets}
