% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longdiff.R
\name{longdiff}
\alias{longdiff}
\title{Differences in Longitudinal Data Function}
\usage{
longdiff(data, ID, timevar, var, timeparam, operator, cutoff)
}
\arguments{
\item{data}{The name of your dataset}

\item{ID}{The variable name corresponding to individual record or subject IDs}

\item{timevar}{The variable corresponding to timepoint (year, month, etc) of your data}

\item{var}{The variable you are interested in looking at differences when compared to baseline}

\item{timeparam}{The timepoint you are interested in comparing to baseline (such as the year)}

\item{operator}{The comparison operator you would like to use to compare the difference to the cutoff, this can be "=", "<", ">", "<=", ">=" or "!="}

\item{cutoff}{The cutoff value (numerical) that you would like to set for your variable of interest}
}
\value{
A vector listing all subject IDs that meet the cutoff criteria for your specified variable
}
\description{
A function that returns the IDs of data in which a condition is met that indicates a positive difference in a variable across two timepoints, for use with longitudinal data in long format
}
\examples{
library(dplyr)
list <- longdiff(diffdata, record_id, timepoint, test_score, 2002, ">", 0)
list
}
\author{
Kayla Esser
}
