% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longdiff.R
\name{longdiff}
\alias{longdiff}
\title{differences in longitudinal data function}
\usage{
longdiff(data, ID, timevar, var, timeparam, cutoff)
}
\arguments{
\item{data}{the name of your dataset}

\item{ID}{the variable name corresponding to individual record or subject IDs}

\item{timevar}{the variable corresponding to timepoint (year, month, etc) of your data}

\item{var}{the variable you are interested in looking at differences within}

\item{timeparam}{the value of the timepoint you are interested in comparing to baseline (such as the year)}

\item{cutoff}{the cutoff value that you would like to set for your variable of interest}
}
\value{
a vector listing all subject IDs that meet the cutoff criteria for your specified variable
}
\description{
a function that returns the IDs of data in which a condition is met that indicates a positive difference in a variable across two timepoints, for use with longitudinal data in long format
}
\examples{
library(dplyr)
list <- longdiff(diffdata, record_id, timepoint, test_score, 2002, 0)
list
}
\author{
Kayla Esser
}
